rename geofence to pathsense
