+++ lib-ng/mapper/pathsense/beacon.js
